refactor: remove /new-home route and redirect /schedule-simulator to /
- Remove src/app/new-home/ directory
- Move HomeScheduleView to components/schedule-simulator/
- Add permanent redirect from /schedule-simulator to /
